Rework STATUS_VIA_WINDOWPORT
This simplifies the core-to-window-port API for updating the status area
by sending all status info every time, instead of piecemeal when
individual status fields change. This makes the status update logic much
easier to reason about, as well as shedding a substantial amount of
book-keeping and generic parameter handling that the old system had to
do.

Because of this API change, all window ports need to be updated to work
with this new system.  This commit has already done this for the TTY
window port, but the others need to be fixed at a later point in time.