Windowport-neutral hitpointbar for NetHack 3.6
This shows a graphical bar over the name and title in the status area
representing remaining hit points, if the "hitpointbar" boolean option
is enabled (it's off by default).  If statuscolors are compiled in,
enabled and set for HP, those colors will be shared for the bar,
otherwise the bar will simply display in inverse.

hitpointbar is designed as a window capability, so any windowport can
support it by including `WC2_HITPOINTBAR` in its windowport flags and
rendering the bar based on `iflags.wc2_hitpointbar`; this commit does
this for TTY when STATUS_VIA_WINDOWPORT is enabled, for example.